Added more comments
Reverted commit that removed mapbox
