--- ShuttleTrackeriOS/ShuttleArrow.swift
@@ -1,82 +1,0 @@
-//
-//  ShuttleArrow.swift
-//  ShuttleTrackeriOS
-//
-//  Created by Andrew Qu on 2/13/19.
-//  Copyright © 2019 WTG. All rights reserved.
-//
-
-import Foundation
-import MapKit
-import UIKit
-
-extension UIImage {
-    
-    func imageWithSize(size:CGSize) -> UIImage
-    {
-        var scaledImageRect = CGRect.zero;
-        
-        let aspectWidth:CGFloat = size.width / self.size.width;
-        let aspectHeight:CGFloat = size.height / self.size.height;
-        let aspectRatio:CGFloat = min(aspectWidth, aspectHeight);
-        
-        scaledImageRect.size.width = self.size.width * aspectRatio;
-        scaledImageRect.size.height = self.size.height * aspectRatio;
-        scaledImageRect.origin.x = (size.width - scaledImageRect.size.width) / 2.0;
-        scaledImageRect.origin.y = (size.height - scaledImageRect.size.height) / 2.0;
-        
-        UIGraphicsBeginImageContextWithOptions(size, false, 0);
-        
-        self.draw(in: scaledImageRect);
-        
-        let scaledImage = UIGraphicsGetImageFromCurrentImageContext();
-        UIGraphicsEndImageContext();
-        
-        return scaledImage!;
-    }
-    
-    func rotate(radians: Float) -> UIImage? {
-        var newSize = CGRect(origin: CGPoint.zero, size: self.size).applying(CGAffineTransform(rotationAngle: CGFloat(radians))).size
-        // Trim off the extremely small float value to prevent core graphics from rounding it up
-        newSize.width = floor(newSize.width)
-        newSize.height = floor(newSize.height)
-        
-        UIGraphicsBeginImageContextWithOptions(newSize, false, self.scale)
-        let context = UIGraphicsGetCurrentContext()!
-        
-        // Move origin to middle
-        context.translateBy(x: newSize.width/2, y: newSize.height/2)
-        // Rotate around middle
-        context.rotate(by: CGFloat(radians))
-        // Draw the image at its center
-        self.draw(in: CGRect(x: -self.size.width/2, y: -self.size.height/2, width: self.size.width, height: self.size.height))
-        
-        let newImage = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        
-        return newImage
-    }
-}
-
-class ShuttleArrow:MKAnnotationView {
-    
-    override var annotation: MKAnnotation? {
-        
-        
-        
-        willSet {
-            guard let shuttle = newValue as? Shuttle else {return}
-            canShowCallout = true
-            calloutOffset = CGPoint(x: -5, y: 5)
-            rightCalloutAccessoryView = UIButton(type: .detailDisclosure)
-            
-            //adjust later for color/rotation
-            image = UIImage(named: shuttle.imageName!)
-            
-            //fix this
-            image = image!.rotate(radians: Float(Float(shuttle.heading-45) * Float(Float.pi/180)))
-//            image = image!.
-            image = image!.imageWithSize(size: CGSize(width: 35, height: 35))
-        }
-    }
-}
